+++ src/clay_board_style_system/core/computed_style.nim
@@ -179,6 +179,7 @@ type
     ckDefault,
     ckPointer,
     ckText,
+    ckCrosshair,
     ckMove,
     ckNotAllowed